add "bit capsule" command, change capsule-list and create-capsule to subcommands of "capsule"
